feat & fix: choose between types of care to receive, refactor code (roughly).
