+++ webapp/src/types/formTypes.ts
@@ -19,10 +19,10 @@ export interface FormData {
   time: string;
   status: string | null;
 
-  // symptoms: string[];
+  careType: string;
   chiefComplaint: SymptomAnswer;
   presentIllness: SymptomAnswer[];
-  inquiries: Omit<AppointmentSymptoms, 'id' | 'appointmentId'>[];
+  inquiries: Omit<AppointmentSymptoms, "id" | "appointmentId">[];
   predicted: boolean;
   prediction: string;
